requirements_out_of_pytorch.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
async-tio==1.3.2
attrs==23.1.0
audioread==3.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chromadb==0.5.0
cn2an==0.5.19
colorama==0.4.6
coloredlogs==15.0.1
Cython==0.29.34
decorator==5.1.1
docx2txt
fastapi==0.115.2
filelock==3.9.0
flatbuffers==23.3.3
frozenlist==1.3.3
fsspec==2023.5.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.4
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
langchain==0.3
langchain-community==0.3
lazy_loader==0.2
librosa==0.10.1
llama-cpp-python==0.3
llvmlite==0.40.0
markdown
MarkupSafe==2.1.2
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
networkx==3.0
numba==0.57.0
numpy==1.24.1
onnxruntime==1.16.1
openai==0.27.6
OpenAIAuth==0.3.6
packaging==23.2
pdfminer-six
Pillow==9.3.0
pooch==1.6.0
proces==0.1.4
prompt-toolkit==3.0.38
protobuf==4.22.4
PyAudio==0.2.13
pycparser==2.21
pypinyin==0.48.0
pyreadline3==3.4.1
PySocks==1.7.1
python-docx
pywin32==306
PyYAML==6.0
regex==2023.5.5
requests==2.28.1
revChatGPT==5.0.0
sentence-transformers==3.2.1
scikit-learn==1.2.2
scipy==1.10.1
sniffio==1.3.0
socksio==1.0.0
soundfile==0.12.1
soxr==0.3.5
sympy==1.12
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.20.3
tqdm==4.65.0
transformers==4.46.2
typeguard==2.13.3
typing_extensions==4.8.0
unstructured
urllib3==1.26.13
wcwidth==0.2.6
WMI==1.5.1
yarl==1.9.2
